Consensus: SQUASHME: MOVEONLY: Consensus::CheckTx() from main.o to consensus/consensus.o
