+++ src/consensus/consensus.h
@@ -1,0 +1,21 @@
+// Copyright (c) 2009-2010 Satoshi Nakamoto
+// Copyright (c) 2009-2014 The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef BITCOIN_CONSENSUS_CONSENSUS_H
+#define BITCOIN_CONSENSUS_CONSENSUS_H
+
+class CTransaction;
+class CValidationState;
+
+/** The maximum allowed size for a serialized block, in bytes (network rule) */
+static const unsigned int MAX_BLOCK_SIZE = 1000000;
+
+namespace Consensus {
+
+bool CheckTx(const CTransaction& tx, CValidationState &state);
+
+} // namespace Consensus
+
+#endif // BITCOIN_CONSENSUS_CONSENSUS_H
