+++ src/merkleblock.cpp
@@ -5,8 +5,8 @@
 
 #include "merkleblock.h"
 
+#include "consensus/consensus.h" // for MAX_BLOCK_SIZE
 #include "hash.h"
-#include "primitives/block.h" // for MAX_BLOCK_SIZE
 #include "utilstrencodings.h"
 
 using namespace std;
